chore: refactor account registry
