+++ crates/app/src/views/onboarding.rs
@@ -1,4 +1,4 @@
-use account::registry::Account;
+use account::login;
 use common::qr::create_qr;
 use gpui::{
     div, img, prelude::FluentBuilder, relative, svg, App, AppContext, ClipboardItem, Context, Div,
@@ -107,7 +107,7 @@ impl Onboarding {
                 cx.spawn(|mut cx| async move {
                     let signer = Arc::new(signer);
 
-                    if Account::login(signer, &cx).await.is_ok() {
+                    if login(signer, &cx).await.is_ok() {
                         _ = cx.update_window(window_handle, |_, window, cx| {
                             window.replace_root(cx, |window, cx| {
                                 Root::new(app::init(window, cx).into(), window, cx)
@@ -149,7 +149,7 @@ impl Onboarding {
         cx.spawn(|this, mut cx| async move {
             let signer = Arc::new(keys);
 
-            if Account::login(signer, &cx).await.is_ok() {
+            if login(signer, &cx).await.is_ok() {
                 _ = cx.update_window(window_handle, |_, window, cx| {
                     window.replace_root(cx, |window, cx| {
                         Root::new(app::init(window, cx).into(), window, cx)
